Moves functions for BART model initialisation
